api-gateway/requirements.txt
email_validator==2.1.1
fastapi==0.110.0
grpcio==1.62.1
grpcio-tools==1.62.1
gunicorn==20.0.4
protobuf==4.25.3
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
uvicorn==0.28.0